### 2021-03-23 #3128 @whedon
[PRE REVIEW]: HyRiver: Hydroclimate Data Retriever
**Submitting author:** @cheginit (<a href="http://orcid.org/0000-0002-5430-6000">Taher Chegini</a>)
**Repository:** <a href="https://github.com/cheginit/HyRiver" target ="_blank">https://github.com/cheginit/HyRiver</a>
**Version:** v0.10.x
