### 2021-03-24 #3128 @cheginit
@kthyng, as @Kevin-Mattheus-Moerman mentioned the above lines of code is only for the collection repo, here's a summary of lines of code (only the main files) for each package using `cloc`:
```bash
❯ for f in $(ls -d1 py*/py*/);do echo $f && cloc $f; done
py3dep/py3dep/
       6 text files.
       4 unique files.
       2 files ignored.

Language                     files          blank        comment           code
Python                           4             75            149            263
SUM:                             4             75            149            263
pydaymet/pydaymet/
       7 text files.
       5 unique files.
       2 files ignored.

Language                     files          blank        comment           code
Python                           5            287            508           1042
SUM:                             5            287            508           1042
pygeohydro/pygeohydro/
      12 text files.
       8 unique files.
       4 files ignored.

Language                     files          blank        comment           code
Python                           8            242            281            865
SUM:                             8            242            281            865
pygeoogc/pygeoogc/
      13 text files.
      10 unique files.
       3 files ignored.

Language                     files          blank        comment           code
Python                           9            531            962           1853
YAML                             1              3              0             25
SUM:                            10            534            962           1878
pygeoutils/pygeoutils/
       7 text files.
       7 unique files.
       0 files ignored.

Language                     files          blank        comment           code
Python                           7            281            369           1037
SUM:                             7            281            369           1037
pynhd/pynhd/
      10 text files.
       6 unique files.
       4 files ignored.

Language                     files          blank        comment           code
Python                           6            252            481            831
SUM:                             6            252            481            831
```

And here's the combined summary:
```bash
❯ for f in $(ls -d1 py*/py*/);do cloc --out $(basename $f).txt $f; done
❯ cloc --sum-reports --out=hyriver.txt *.txt
❯ cat hyriver.txt.lang
github.com/AlDanial/cloc v 1.89
Language                     files          blank        comment           code
Python                          39           1668           2750           5891
YAML                             1              3              0             25
SUM:                            40           1671           2750           5916
```

Please let me know if more info is needed.
