### 2021-03-27 c737b17
ENH: Remove unnecessary hooks
