### 2021-03-28 ec1e407
DOC: Use JupyterLab for the binder instead of notebook [skip ci]
