### 2021-04-02 #3128 @kthyng
@evanleeturner I hope things are going well with you! Are you interested in reviewing this submission to JOSS? As a reminder, you can read about the process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and we are asking for reviews to be finished within about 4-6 weeks but since it is often an iterative process, you would want to start sooner than that. Thanks for your consideration.
