### 2021-04-02 #3128 @kthyng
Hi @kmarkert are you interested in reviewing this submission for the Journal of Open Source Software? You can read about the process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) and we are asking for reviews to be finished within about 4-6 weeks but since it is often an iterative process, you would want to start sooner than that. Thanks for your consideration.
