### 2021-04-06 ba9ee65
ENH: Use nbsphinx_link extension to get the example notebooks from the newly added submodule
