+++ docs/source/conf.py
@@ -45,6 +45,7 @@ extensions = [
     "IPython.sphinxext.ipython_directive",
     "IPython.sphinxext.ipython_console_highlighting",
     "nbsphinx",
+    "nbsphinx_link",
     "sphinx_gallery.load_style",
     "sphinx.ext.mathjax",
     "recommonmark",
