+++ docs/source/docs/source/examples/nhdplus.py
@@ -1,260 +1,0 @@
-# ---
-# jupyter:
-#   jupytext:
-#     text_representation:
-#       extension: .py
-#       format_name: sphinx
-#       format_version: '1.1'
-#       jupytext_version: 1.10.3
-#   kernelspec:
-#     display_name: Python 3
-#     language: python
-#     name: python3
-# ---
-
-"""
-# NHDPlus and Accumulation
-"""
-
-import cmocean.cm as cmo
-import matplotlib.patches as mpatches
-import matplotlib.pyplot as plt
-import pynhd as nhd
-from pynhd import NLDI, NHDPlusHR, WaterData
-
-""
-import warnings
-
-warnings.filterwarnings("ignore")
-
-###############################################################################
-# [PyNHD](https://github.com/cheginit/pynhd) provides access to the Hydro Network-Linked Data Index ([NLDI](https://labs.waterdata.usgs.gov/about-nldi/index.html)) and the [WaterData](https://labs.waterdata.usgs.gov/geoserver/web/wicket/bookmarkable/org.geoserver.web.demo.MapPreviewPage?1) web services for navigating and subsetting [NHDPlus](https://nhdplus.com/NHDPlus) V2 database. Additionally, you can download NHDPlus High Resolution data as well.
-#
-# First, let's get the watershed geometry of the contributing basin of a USGS station using `NLDI`:
-
-nldi = NLDI()
-station_id = "01031500"
-
-basin = nldi.get_basins(station_id)
-
-###############################################################################
-# The `navigate_byid` class method can be used to navigate NHDPlus in both upstream and downstream of any point in the database. The available feature sources are ``comid``, ``huc12pp``, ``nwissite``, ``wade``, ``wqp``. Let's get ComIDs and flowlines of the tributaries and the main river channel in the upstream of the station.
-
-flw_main = nldi.navigate_byid(
-    fsource="nwissite",
-    fid=f"USGS-{station_id}",
-    navigation="upstreamMain",
-    source="flowlines",
-    distance=1000,
-)
-
-flw_trib = nldi.navigate_byid(
-    fsource="nwissite",
-    fid=f"USGS-{station_id}",
-    navigation="upstreamTributaries",
-    source="flowlines",
-    distance=1000,
-)
-
-###############################################################################
-# We can get other USGS stations upstream (or downstream) of the station and even set a distance limit (in km):
-
-st_all = nldi.navigate_byid(
-    fsource="nwissite",
-    fid=f"USGS-{station_id}",
-    navigation="upstreamTributaries",
-    source="nwissite",
-    distance=1000,
-)
-
-st_d20 = nldi.navigate_byid(
-    fsource="nwissite",
-    fid=f"USGS-{station_id}",
-    navigation="upstreamTributaries",
-    source="nwissite",
-    distance=20,
-)
-
-###############################################################################
-# Now, let's get the [HUC12 pour points](https://www.sciencebase.gov/catalog/item/5762b664e4b07657d19a71ea):
-
-pp = nldi.navigate_byid(
-    fsource="nwissite",
-    fid=f"USGS-{station_id}",
-    navigation="upstreamTributaries",
-    source="huc12pp",
-    distance=1000,
-)
-
-###############################################################################
-# Let's plot the vector data:
-
-ax = basin.plot(facecolor="none", edgecolor="k", figsize=(8, 8))
-st_all.plot(
[…]
