### 2021-05-15 0d7ff16
ENH: Total overhaul of the website using sphinx-book-theme
