+++ docs/source/conf.py
@@ -1,15 +1,3 @@
-# Configuration file for the Sphinx documentation builder.
-#
-# This file only contains a selection of the most common options. For a full
-# list see the documentation:
-# https://www.sphinx-doc.org/en/master/usage/configuration.html
-
-# -- Path setup --------------------------------------------------------------
-
-# If extensions (or modules to document with autodoc) are in another directory,
-# add these directories to sys.path here. If the directory is relative to the
-# documentation root, use os.path.abspath to make it absolute, like shown here.
-#
 import datetime
 import os
 import sys
@@ -18,7 +6,6 @@ sys.path.insert(0, os.path.abspath("../.."))
 
 
 # -- Project information -----------------------------------------------------
-
 project = "HyRiver"
 author = "Taher Chegini"
 copyright = f"2019-{datetime.datetime.now().year}, {author}"
@@ -34,27 +21,29 @@ release = version
 
 # -- General configuration ---------------------------------------------------
 
-# Add any Sphinx extension module names here, as strings. They can be
-# extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
-# ones.
 extensions = [
     "autoapi.extension",
-    "sphinx.ext.autodoc.typehints",
     "sphinx.ext.napoleon",
     "sphinx.ext.intersphinx",
+    "sphinx.ext.extlinks",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.mathjax",
     "IPython.sphinxext.ipython_directive",
     "IPython.sphinxext.ipython_console_highlighting",
     "nbsphinx",
     "nbsphinx_link",
     "sphinx_gallery.load_style",
-    "sphinx.ext.mathjax",
     "recommonmark",
+    "sphinx_panels",
+    "sphinxext.opengraph",
     "sphinx_copybutton",
 ]
 
+# intersphinx configurations
 intersphinx_mapping = {
     "python": ("https://docs.python.org/3/", None),
     "xarray": ("https://xarray.pydata.org/en/stable/", None),
+    "dask": ("https://docs.dask.org/en/latest/", None),
     "geopandas": ("https://geopandas.org/", None),
     "pandas": ("https://pandas.pydata.org/pandas-docs/stable/", None),
     "numpy": ("https://numpy.org/doc/stable/", None),
@@ -63,6 +52,7 @@ intersphinx_mapping = {
     "folium": ("https://python-visualization.github.io/folium/", None),
 }
 
+# autoapi configurations
 autoapi_dirs = [
     "../../pygeoogc",
     "../../pygeoutils",
@@ -71,18 +61,18 @@ autoapi_dirs = [
     "../../py3dep",
     "../../pydaymet",
 ]
-autoapi_ignore = ["**ipynb_checkpoints**", "**tests**", "**setup**", "**generate_pip**", "**conf**"]
-autoapi_options = [
-    "members",
-    "inherited-members",
-    "undoc-members",
-    # 'private-members',
-    "show-inheritance",
-    # 'show-module-summary',
-    # 'special-members',
-    # 'imported-members',
+autoapi_ignore = [
+    "**ipynb_checkpoints**",
+    "**tests**",
+    "**setup**",
+    "**conf**",
+    "**exceptions**",
+    "**print_versions**",
 ]
-autodoc_typehints = "description"
+autoapi_options = ["members"]
+autoapi_member_order = "groupwise"
+autoapi_keep_files = True
+autoapi_add_toctree_entry = False
 modindex_common_prefix = [
     "pygeoogc.",
     "pygeoutils.",
@@ -92,6 +82,10 @@ modindex_common_prefix = [
[…]
