+++ docs/source/index.rst
@@ -1,31 +1,33 @@
-.. image:: https://raw.githubusercontent.com/cheginit/HyRiver-examples/main/notebooks/_static/hyriver_logo_text.png
-    :target: https://github.com/cheginit/HyRiver-examples
+.. |async| image:: https://img.shields.io/pypi/v/async_retriever?label=AsyncRetriever&color=green
+    :target: https://github.com/cheginit/async_retriever
+    :alt: PyPi Version
 
-|
+.. |pygeohydro| image:: https://img.shields.io/pypi/v/pygeohydro?label=PyGeoHydro&color=green
+    :target: https://github.com/cheginit/pygeohydro
+    :alt: PyPi Version
 
-.. |pygeohydro| image:: https://github.com/cheginit/pygeohydro/actions/workflows/test.yml/badge.svg
-    :target: https://github.com/cheginit/pygeohydro/actions/workflows/test.yml
-    :alt: Github Actions
+.. |pygeoogc| image:: https://img.shields.io/pypi/v/pygeoogc?label=PyGeoOGC&color=green
+    :target: https://github.com/cheginit/pygeoogc
+    :alt: PyPi Version
 
-.. |pygeoogc| image:: https://github.com/cheginit/pygeoogc/actions/workflows/test.yml/badge.svg
-    :target: https://github.com/cheginit/pygeoogc/actions/workflows/test.yml
-    :alt: Github Actions
+.. |pygeoutils| image:: https://img.shields.io/pypi/v/pygeoutils?label=PyGeoUtils&color=green
+    :target: https://github.com/cheginit/pygeoutils
+    :alt: PyPi Version
 
-.. |pygeoutils| image:: https://github.com/cheginit/pygeoutils/actions/workflows/test.yml/badge.svg
-    :target: https://github.com/cheginit/pygeoutils/actions/workflows/test.yml
-    :alt: Github Actions
+.. |pynhd| image:: https://img.shields.io/pypi/v/pynhd?label=PyNHD&color=green
+    :target: https://github.com/cheginit/pynhd
+    :alt: PyPi Version
 
-.. |pynhd| image:: https://github.com/cheginit/pynhd/actions/workflows/test.yml/badge.svg
-    :target: https://github.com/cheginit/pynhd/actions/workflows/test.yml
-    :alt: Github Actions
+.. |py3dep| image:: https://img.shields.io/pypi/v/py3dep?label=Py3DEP&color=green
+    :target: https://github.com/cheginit/py3dep
+    :alt: PyPi Version
 
-.. |py3dep| image:: https://github.com/cheginit/py3dep/actions/workflows/test.yml/badge.svg
-    :target: https://github.com/cheginit/py3dep/actions/workflows/test.yml
-    :alt: Github Actions
+.. |pydaymet| image:: https://img.shields.io/pypi/v/pydaymet?label=PyDaymet&color=green
+    :target: https://github.com/cheginit/pydaymet
+    :alt: PyPi Version
 
-.. |pydaymet| image:: https://github.com/cheginit/pydaymet/actions/workflows/test.yml/badge.svg
-    :target: https://github.com/cheginit/pydaymet/actions/workflows/test.yml
-    :alt: Github Actions
+.. |async_stat| image:: https://pepy.tech/badge/async_retriever
+    :target: https://pepy.tech/project/async_retriever
 
 .. |pygeohydro_stat| image:: https://pepy.tech/badge/hydrodata
     :target: https://pepy.tech/project/hydrodata
@@ -51,12 +53,9 @@
     :target: https://pepy.tech/project/pydaymet
     :alt: Download Stat
 
[…]
