+++ docs/source/autoapi/pydaymet/index.rst
@@ -1,0 +1,20 @@
+:mod:`pydaymet`
+===============
+
+.. py:module:: pydaymet
+
+.. autoapi-nested-parse::
+
+   Top-level package for PyDaymet.
+
+
+
+Submodules
+----------
+.. toctree::
+   :titlesonly:
+   :maxdepth: 1
+
+   pydaymet/index.rst
+
+
