+++ docs/source/autoapi/index.rst
@@ -1,0 +1,49 @@
+Reference
+=========
+
+Versions
+--------
+
+- |pygeoogc|
+- |pygeoutils|
+- |pynhd|
+- |pygeohydro|
+- |py3dep|
+- |pydaymet|
+
+APIs
+----
+
+.. toctree::
+   :titlesonly:
+
+   /autoapi/pygeoogc/index
+   /autoapi/pygeoutils/index
+   /autoapi/pynhd/index
+   /autoapi/py3dep/index
+   /autoapi/pygeohydro/index
+   /autoapi/pydaymet/index
+
+.. |pygeohydro| image:: https://img.shields.io/pypi/v/pygeohydro?label=PyGeoHydro&color=blue
+    :target: https://github.com/cheginit/pygeohydro
+    :alt: PyPi Version
+
+.. |pygeoogc| image:: https://img.shields.io/pypi/v/pygeoogc?label=PyGeoOGC&color=blue
+    :target: https://github.com/cheginit/pygeoogc
+    :alt: PyPi Version
+
+.. |pygeoutils| image:: https://img.shields.io/pypi/v/pygeoutils?label=PyGeoUtils&color=blue
+    :target: https://github.com/cheginit/pygeoutils
+    :alt: PyPi Version
+
+.. |pynhd| image:: https://img.shields.io/pypi/v/pynhd?label=PyNHD&color=blue
+    :target: https://github.com/cheginit/pynhd
+    :alt: PyPi Version
+
+.. |py3dep| image:: https://img.shields.io/pypi/v/py3dep?label=Py3DEP&color=blue
+    :target: https://github.com/cheginit/py3dep
+    :alt: PyPi Version
+
+.. |pydaymet| image:: https://img.shields.io/pypi/v/pydaymet?label=PyDaymet&color=blue
+    :target: https://github.com/cheginit/pydaymet
+    :alt: PyPi Version

[…]
