### 2021-05-17 #3175 @raoulcollenteur
Started on the Review today, interesting work! I will edit this comment over the next couple of days.

**Summary**
HyRiver is a collection (software stack) of six python packages that aim to improve the workflow of data collection and processing for hydrological and climatological data from public web services. This is an important and often time-consuming step in hydrological studies, and I congratulate the Authors on their work. Although several alternatives are available (also listed in the paper), this stack seems to provide a unified access to many data sources in the US. The stack is subdivided over several (8?) GitHub repositories, one of which combines all the packages (https://github.com/cheginit/HyRiver). The package is currently actively maintained and looking at some of the user statistics there is a clear demand for this package. The software stack is a significant scholarly contribution in my opinion and suitable for publication in JOSS.

**Package Organization**
My only real concern is the dispersion of the information and code. The packages, information, and support structures are subdivided over several GitHub repositories. I think it would be good to (at least) put all of these under a single GitHub organization and try to reduce the dispersion of information a bit. Documentation is already provided from a single ReadTheDocs website, which is a nice entry point for new users. It could help to do the same for the user support structures (e.g., Code Issues, User questions), which now seem spread over all repositories. Maybe the Authors prefer it the way it is, but I can imagine it is more work to maintain and keep track of user discussions.

Would it not be a better idea to combine all the packages into a single Python packages and have the current packages as sub-packages of hyriver? Given the limited amount of code contained in the individual packages I think this would not be a bad option and it would probably reduce the maintenance cost in the long term. Is there a specific reason different packages were created instead of a single one? A related question I have is if it is possible to use the packages independently. If so, this would be worth mentioning in the docs installation guide. If any case, it could be considered to put all packages into a single python package and use sub-packages to structure.

Here's a list of minor suggestions to improve the package, JOSS paper and documentation (may I add more later):

- [ ] I think most of the packages are only applicable in certain geographic areas (US?), this should be made explicit in the Docs and JOSS paper.
- [ ] Consider adding a page with the features of each package to the documentation website (hyriver.readthedocs.io)
- [ ] The Contributing page seems outdates and can be updated: https://hyriver.readthedocs.io/en/latest/contributing.html
- [ ] I could not easily find a location where user can "seek support", maybe using a central GitHub Discussion or Slack is an idea?

I am still trying to run the examples from https://github.com/cheginit/HyRiver-examples but will need some more time for that. I get lots of errors when trying to import packages, will post some here later.

- [ ] import pydaymet: AttributeError: 'LGEOS360' object has no attribute 'GEOSBufferWithParams'
- [ ] import pygeohydro: AttributeError: module 'pyproj' has no attribute 'crs'
- [ ] import py3dep: AttributeError: module 'pyproj' has no attribute 'crs'

Great work!

Cheers,
Raoul
