### 2021-05-17 #3175 @cheginit
@raoulcollenteur Thanks for accepting to review this submission and your comments.

> My only real concern is the dispersion of the information and code. The packages, information, and support structures are subdivided over several GitHub repositories. I think it would be good to (at least) put all of these under a single GitHub organization and try to reduce the dispersion of information a bit. Documentation is already provided from a single ReadTheDocs website, which is a nice entry point for new users. It could help to do the same for the user support structures (e.g., Code Issues, User questions), which now seem spread over all repositories. Maybe the Authors prefer it the way it is, but I can imagine it is more work to maintain and keep track of user discussions.

> Would it not be a better idea to combine all the packages into a single Python packages and have the current packages as sub-packages of hyriver? Given the limited amount of code contained in the individual packages I think this would not be a bad option and it would probably reduce the maintenance cost in the long term. Is there a specific reason different packages were created instead of a single one?

The software stack was initially a single package called `hydrodata`. As the package started to grow and I added support for more service and new functionalities, the maintenance became more difficult. So I decided to break it down into six self-contained packages, each of which has a limited scope and serves different target users. For example, some users might only want to have access to river network data, so instead of installing a hefty package with many dependencies and irrelevant functions, they can just install PyNHD that offers what they need. Also, from development point of view (for the core developers as well as contributors), working on new functionalities and tracking down issues become easier as the code base is smaller for each package and better organized. The only downside is more work for releasing six packages instead of one which is not big of an issue overall.

Moreover, the packages are divided into two categories: Low- and high-level APIs. PyGeoOGC and PyGeoUtils (and the recently added package called [`async_retriever`](https://github.com/cheginit/async_retriever)) are the engines, if you will, that the other packages use to connect to specific web services. The other packages are specific to the public web services that are provided by US agencies.

> A related question I have is if it is possible to use the packages independently. If so, this would be worth mentioning in the docs installation guide. If any case, it could be considered to put all packages into a single python package and use sub-packages to structure.

Yes, the packages are standalone. The README of each one of the repositories contains descriptions, instructions, and examples that are specific to that package. I will modify the paper as well as repositories' README to reflect this explicitly.

> I think most of the packages are only applicable in certain geographic areas (US?), this should be made explicit in the Docs and JOSS paper.

Yes, currently only US is supported but contributions are welcome as the low-level packages are generic and can be used to access any of the supported web service protocols. I will make this more explicit in the paper and repositories.

> Consider adding a page with the features of each package to the documentation website (hyriver.readthedocs.io)

I thought about this when I was designing the website. Since each package has a repository with a README that has all the relevant information, instead of repeating the same thing in the website, I decided to include a one-line description of each package and provide the link to each repository. Additionally, the example gallery includes most of the things that are in READMEs and even more, so in my opinion including them again in separate pages will be repetitive.

> The Contributing page seems outdates and can be updated: [hyriver.readthedocs.io/en/latest/contributing.html](https://hyriver.readthedocs.io/en/latest/contributing.html)

I will take care of it.

> I could not easily find a location where user can "seek support", maybe using a central GitHub Discussion or Slack is an idea?

The idea is that users can open issues for packages in their corresponding repositories or use the main repository (HyRiver). Users have been using both. I just added a [discussions](https://github.com/cheginit/HyRiver/discussions) tab to the HyRiver's repository.

> I am still trying to run the examples from [cheginit/HyRiver-examples](https://github.com/cheginit/HyRiver-examples) but will need some more time for that. I get lots of errors when trying to import packages, will post some here later.
>
>     * [ ]  import pydaymet: AttributeError: 'LGEOS360' object has no attribute 'GEOSBufferWithParams'
>
>     * [ ]  import pygeohydro: AttributeError: module 'pyproj' has no attribute 'crs'
>
>     * [ ]  import py3dep: AttributeError: module 'pyproj' has no attribute 'crs'

It would be great and appreciated if you can provide more details and, perhaps, open an issue so I can reproduce and fix them.

By the way, I am working on a new release as users reported some issues after `xarray` released a new version recently that included some breaking changes. I have already took care of the issues in the main branches of the affected repositories and will release a new version soon.
