+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
   - repo: git://github.com/pre-commit/pre-commit-hooks
-    rev: v4.0.0
+    rev: v4.0.1
     hooks:
     - id: trailing-whitespace
     - id: check-executables-have-shebangs
@@ -25,4 +25,11 @@ repos:
     - id: blacken-docs
       name: Autoformat codes in docstrings with blacken-docs
       additional_dependencies: [black]
-      args: [-t, py38, -l, "88"]
+      args: [-t, py38, -l, "100"]
+
+  - repo: https://github.com/codespell-project/codespell
+    rev: v2.0.0
+    hooks:
+    - id: codespell
+      name: Checks for common misspellings in text files.
+
