### 2021-05-27 0e5b78d
DOC: Update to reflect the current state of the packages openjournals/joss-reviews#3175
