+++ docs/source/changelogs/pygeohydro.rst
@@ -1,0 +1,1 @@
+.. include:: ../../../pygeohydro/HISTORY.rst
