+++ HISTORY.rst
@@ -1,327 +1,0 @@
-=======
-History
-=======
-
-0.10.2 (2021-03-27)
--------------------
-
-- Add annoucement regarding the new name for the softwate stack, HyRiver.
-- Improve ``pip`` installation and release workflow.
-
-PyGeoUtils
-~~~~~~~~~~
-- Setting transform of the merged dataset explicitly to address cheginit/py3dep#3
-
-0.10.1 (2021-03-06)
--------------------
-
-PyGeoHydro
-~~~~~~~~~~
-- Add ``lxml`` to deps.
-
-0.10.0 (2021-03-06)
--------------------
-
-PyGeoHydro
-~~~~~~~~~~
-- The official first release of PyGeoHydro with a new name and logo.
-- Replace cElementTree with ElementTree since it's been deprecated by ``defusedxml``.
-- Make ``mypy`` checks more strict and fix all the errors and prevent possible
-  bugs.
-- Speed up CI testing by using ``mamba`` and caching.
-
-
-0.9.2 (2021-03-02)
-------------------
-
-PyGeoHydro
-~~~~~~~~~~
-- Rename ``hydrodata`` package to ``pygeohydro`` for publication on JOSS.
-- In ``NWIS.get_info``, drop rows that don't have mean daily discharge data instead of slicing.
-- Add support for the National Inventory of Dams (NID) via ``get_nid`` function.
-- Speed up Github Actions by using ``mamba`` and caching.
-- Improve ``pip`` installation by adding ``pyproject.toml``.
-
-0.9.1 (2021-02-22)
-------------------
-
-PyGeoHydro
-~~~~~~~~~~
-- Fix an issue with ``NWIS.get_info`` method where stations with False values as their
-  ``hcdn_2009`` value were returned as ``None`` instead.
-
-0.9.0 (2021-02-14)
-------------------
-
-- Bump versions of packages across the stack to the same version.
-
-PyGeoHydro
-~~~~~~~~~~
-- Use the new pynhd function for getting basins, ``NLDI.get_basisn``.
-- Made ``mypy`` checks more strict and added all the missing type annotations.
-
-PyGeoOGC
-~~~~~~~~
-- Add support for query by point and multi-points to ``ArcGISRESTful.bygeom``.
-- Add support for buffer distance to ``ArcGISRESTful.bygeom``.
-- Add support for generating ESRI-based queries for points and multi-points
-  to ``ESRIGeomQuery``.
-- Add all the missing type annotations.
-- Update the Daymet url to version 4. You can check the release information
-  `here <https://daac.ornl.gov/DAYMET/guides/Daymet_Daily_V4.html>`_
-- Use ``cytoolz`` library for some of the operations for improving performance.
-- Add ``extent`` property to ``ArcGISRESTful`` class that get the spatial extent
-  of the service.
-- Add url to ``airmap`` service for getting elevation data at 30 m resolution.
-
-PyGeoUtils
-~~~~~~~~~~
-- Add ``gtiff2file`` for saving raster responses as ``geotiff`` file(s).
-- Fix an error in ``_get_nodata_crs`` for handling nodata value when its value in the source
-  is None.
-- Fix the warning during the ``GeoDataFrame`` generation in ``json2geodf`` when there is
-  no geometry column in the input json.
-
-PyNHD
-~~~~~
-- Add a new function for getting basins geometries for a list of USGS station IDs.
-  The function is a method of ``NLDI`` class called ``get_basins``. So, now
-  ``NLDI.getfeature_byid`` function does not have a basin flag. This change
-  makes getting geometries easier and faster.
-- Remove ``characteristics_dataframe`` method from ``NLDI`` and made a standalone function
[…]
