+++ docs/source/changelogs/pydaymet.rst
@@ -1,0 +1,1 @@
+.. include:: ../../../pydaymet/HISTORY.rst
