+++ docs/source/history.rst
@@ -1,1 +1,58 @@
-.. include:: ../../HISTORY.rst
+Changelogs
+==========
+
+.. panels::
+    :container: container-lg pb-3
+    :body: text-center
+    :column: col-lg-3 col-md-3 col-sm-4 col-xs-6 p-2
+
+    .. link-button:: changelogs/pynhd
+        :type: ref
+        :text: PyNHD
+        :classes: stretched-link
+
+    ---
+
+    .. link-button:: changelogs/pygeohydro
+        :type: ref
+        :text: PyGeoHydro
+        :classes: stretched-link
+
+    ---
+
+    .. link-button:: changelogs/py3dep
+        :type: ref
+        :text: Py3DEP
+        :classes: stretched-link
+
+    ---
+
+    .. link-button:: changelogs/pydaymet
+        :type: ref
+        :text: PyDaymet
+        :classes: stretched-link
+
+    ---
+
+    .. link-button:: changelogs/pygeoogc
+        :type: ref
+        :text: PyGeoOGC
+        :classes: stretched-link
+
+    ---
+
+    .. link-button:: changelogs/pygeoutils
+        :type: ref
+        :text: PyGeoUtils
+        :classes: stretched-link
+
+.. toctree::
+    :maxdepth: 1
+    :hidden:
+
+    PyGeoOGC <changelogs/pygeoogc>
+    PyGeoUtils <changelogs/pygeoutils>
+    PyNHD <changelogs/pynhd>
+    Py3DEP <changelogs/py3dep>
+    PyGeoHydro <changelogs/pygeohydro>
+    PyDaymet <changelogs/pydaymet>
