### 2021-05-27 a177f5e
DOC: Reorder packages following the order on the website
