### 2021-05-27 c484af3
DOC: Add async_retriever and more example notebooks openjournals/joss-reviews#3175
