+++ docs/source/autoapi/py3dep/index.rst
@@ -15,6 +15,7 @@ Submodules
    :titlesonly:
    :maxdepth: 1
 
+   cli/index.rst
    py3dep/index.rst
 
 
