### 2021-06-11 6cfee36
DOC: Add link to pangeo showcase presentation
