+++ docs/source/index.rst
@@ -72,9 +72,14 @@ Hydroclimate Data Retriever
 |
 
 `HyRiver <https://hyriver.readthedocs.io>`__ (formerly named
-`hydrodata <https://pypi.org/project/hydrodata>`__) is a software stack consisting of seven
-Python libraries that are designed to aid in watershed analysis through web services.
-Currently, this project only includes hydrology and climatology data within the US.
+`hydrodata <https://pypi.org/project/hydrodata>`__) is a software stack
+consisting of seven Python libraries that are designed to aid in watershed
+analysis through web services. Currently, this project only includes hydrology
+and climatology data within the US.
+
+You can watch this videos for a quick overview of ``HyRiver``:
+
+* `Pangeo Showcase <https://discourse.pangeo.io/t/may-26-2021-accessing-hydrology-and-climatology-database-using-web-services-through-python/1521>`__
 
 .. toctree::
     :maxdepth: 1
