### 2021-06-11 10b13ac
ENH: Add a new example and remake api
