+++ docs/source/autoapi/noxfile/index.rst
@@ -1,0 +1,8 @@
+:orphan:
+
+:mod:`noxfile`
+==============
+
+.. py:module:: noxfile
+
+
