### 2021-07-28 #3175 @raoulcollenteur
Sorry @kthyng , the email indeed helped!

Thanks @cheginit for all the changes to the code, documentation, and the JOSS paper. I have now checked most boxes in the Review form. Just one issue that remains I think, about running the examples after a clean installation.

> > ```
> > * [ ]  import pydaymet: AttributeError: 'LGEOS360' object has no attribute 'GEOSBufferWithParams'
> >
> > * [ ]  import pygeohydro: AttributeError: module 'pyproj' has no attribute 'crs'
> >
> > * [ ]  import py3dep: AttributeError: module 'pyproj' has no attribute 'crs'
> > ```
>
> It would be great and appreciated if you can provide more details and, perhaps, open an issue so I can reproduce and fix them.

I checked, and these problems seem to occur because there are no minimum dependencies defined (e.g., numpy > 1.6) for installation. My environment had Pyproj 1.9.6, which causes errors. The same holds for geopandas (I had 0.6). Most examples won't work in this case, so it seems necessary to add minimum versions for all dependencies during installation throughout the project. I suggest to create a clean environment and [run all the examples](https://github.com/cheginit/HyRiver-Examples/tree/5ca8081202867d924bec83635583c55795fc6ab6), and update the minimum versions where necessary. I now did this manually for my environment and could run most of the examples already.

Cheers,
Raoul
