### 2021-07-28 #3175 @cheginit
@raoulcollenteur Thanks for the feedback, it's a good idea. I added minimum versions as you suggested, hopefully it will resolve the issue.

Some of the example won't work since two of the web services that I am using are having technical issues that their developers are working on to fix. I have already open two issues (cheginit/pynhd#17 and cheginit/pygeohydro#51) to update the status of those web services.
