+++ docs/source/autoapi/pygeoogc/pygeoogc/index.rst
@@ -32,15 +32,6 @@ Module Contents
                   simultaneously and will return the requests partially. It's recommended
                   to avoid performing threading unless you are certain the web service can handle it.
 
-   .. method:: get_features(self, return_m: bool = False) -> List[Dict[str, Any]]
-
-      Get features based on the feature IDs.
-
-      :Parameters: **return_m** (:class:`bool`) -- Whether to activate the Return M (measure) in the request, defaults to False.
-
-      :returns: :class:`dict` -- (Geo)json response from the web service.
-
-
    .. method:: oids_byfield(self, field: str, ids: Union[str, List[str]]) -> None
 
       Get Object IDs based on a list of field IDs.
@@ -53,11 +44,9 @@ Module Contents
 
       Get feature IDs within a geometry that can be combined with a SQL where clause.
 
-      :Parameters: * **geom** (:class:`LineString`, :class:`Polygon`, :class:`Point`, :class:`MultiPoint`, :class:`tuple`, or :class:`list` of :class:`tuples`) --
-
-                     A geometry (LineString, Polygon, Point, MultiPoint), tuple of length
-                      2 (``(x, y)``), a list of tuples of length 2 (``[(x, y), ...]``), or bounding box
-                      (tuple of length 4 (``(xmin, ymin, xmax, ymax)``)).
+      :Parameters: * **geom** (:class:`LineString`, :class:`Polygon`, :class:`Point`, :class:`MultiPoint`, :class:`tuple`, or :class:`list` of :class:`tuples`) -- A geometry (LineString, Polygon, Point, MultiPoint), tuple of length two
+                     (``(x, y)``), a list of tuples of length 2 (``[(x, y), ...]``), or bounding box
+                     (tuple of length 4 (``(xmin, ymin, xmax, ymax)``)).
                    * **geo_crs** (:class:`str`) -- The spatial reference of the input geometry, defaults to EPSG:4326.
                    * **spatial_relation** (:class:`str`, *optional*) -- The spatial relationship to be applied on the input geometry
                      while performing the query. If not correct a list of available options is shown.
@@ -72,8 +61,8 @@ Module Contents
                      * ``esriSpatialRelTouches``
                      * ``esriSpatialRelWithin``
                      * ``esriSpatialRelRelation``
-                   * **sql_clause** (:class:`str`, *optional*) -- A valid SQL 92 WHERE clause, default to None.
-                   * **distance** (:class:`int`, *optional*) -- The buffer distance in meters for the input geometries in meters, default to None.
+                   * **sql_clause** (:class:`str`, *optional*) -- Valid SQL 92 WHERE clause, default to None.
+                   * **distance** (:class:`int`, *optional*) -- Buffer distance in meters for the input geometries, default to None.
 
 
    .. method:: oids_bysql(self, sql_clause: str) -> None
@@ -132,7 +121,7 @@ Module Contents
 
                   The data format to request for data from the service, defaults to None which
                    throws an error and includes all the available format offered by the service.
-                * **version** (:class:`str`, *optional*) -- The WFS service version which should be either 1.1.1, 1.3.0, or 2.0.0.
+                * **version** (:class:`str`, *optional*) -- The WFS service version which should be either 1.0.0, 1.1.0, or 2.0.0.
                   Defaults to 2.0.0.
                 * **crs** (:class:`str`, *optional*) -- The spatial reference system to be used for requesting the data, defaults to
                   epsg:4326.
@@ -140,7 +129,7 @@ Module Contents
                   to False if you are sure all the WFS settings such as layer and crs are correct
[…]
