+++ docs/source/autoapi/pygeoutils/pygeoutils/index.rst
@@ -12,63 +12,6 @@
 Module Contents
 ---------------
 
-.. py:class:: MatchCRS(in_crs: str, out_crs: str)
-
-   Reproject a geometry to another CRS.
-
-   :Parameters: * **in_crs** (:class:`str`) -- Spatial reference of the input geometry
-                * **out_crs** (:class:`str`) -- Target spatial reference
-
-   .. method:: bounds(self, geom: Tuple[float, float, float, float]) -> Tuple[float, float, float, float]
-
-      Reproject a bounding box to the specified output CRS.
-
-      :Parameters: **geom** (:class:`tuple`) -- Input bounding box (xmin, ymin, xmax, ymax).
-
-      :returns: :class:`tuple` -- Input bounding box in the specified CRS.
-
-      .. rubric:: Examples
-
-      >>> from pygeoogc import MatchCRS
-      >>> bbox = (-7766049.665, 5691929.739, -7763049.665, 5696929.739)
-      >>> MatchCRS("epsg:3857", "epsg:4326").bounds(bbox)
-      (-69.7636111130079, 45.44549114818127, -69.73666165448431, 45.47699468552394)
-
-
-   .. method:: coords(self, geom: List[Tuple[float, float]]) -> List[Tuple[Any, ...]]
-
-      Reproject a list of coordinates to the specified output CRS.
-
-      :Parameters: **geom** (:class:`list` of :class:`tuple`) -- Input coords [(x1, y1), ...].
-
-      :returns: :class:`tuple` -- Input list of coords in the specified CRS.
-
-      .. rubric:: Examples
-
-      >>> from pygeoogc import MatchCRS
-      >>> coords = [(-7766049.665, 5691929.739)]
-      >>> MatchCRS("epsg:3857", "epsg:4326").coords(coords)
-      [(-69.7636111130079, 45.44549114818127)]
-
-
-   .. method:: geometry(self, geom: Union[Polygon, LineString, MultiLineString, MultiPolygon, Point, MultiPoint]) -> Union[Polygon, LineString, MultiLineString, MultiPolygon, Point, MultiPoint]
-
-      Reproject a geometry to the specified output CRS.
-
-      :Parameters: **geom** (:class:`LineString`, :class:`MultiLineString`, :class:`Polygon`, :class:`MultiPolygon`, :class:`Point`, or :class:`MultiPoint`) -- Input geometry.
-
-      :returns: :class:`LineString`, :class:`MultiLineString`, :class:`Polygon`, :class:`MultiPolygon`, :class:`Point`, or :class:`MultiPoint` -- Input geometry in the specified CRS.
-
-      .. rubric:: Examples
-
-      >>> from pygeoogc import MatchCRS
-      >>> from shapely.geometry import Point
-      >>> point = Point(-7766049.665, 5691929.739)
-      >>> MatchCRS("epsg:3857", "epsg:4326").geometry(point).xy
-      (array('d', [-69.7636111130079]), array('d', [45.44549114818127]))
-
-
-
 .. function:: arcgis2geojson(arcgis: Union[str, Dict[str, Any]], id_attr: Optional[str] = None) -> Dict[str, Any]
 
    Convert ESRIGeoJSON format to GeoJSON.
@@ -83,7 +26,28 @@ Module Contents
    :returns: :class:`dict` -- A GeoJSON file readable by GeoPandas.
 
 
-.. function:: gtiff2xarray(r_dict: Dict[str, bytes], geometry: Union[Polygon, MultiPolygon, Tuple[float, float, float, float]], geo_crs: str, ds_dims: Tuple[str, str] = ('y', 'x'), driver: str = 'GTiff') -> Union[xr.DataArray, xr.Dataset]
[…]
