### 2021-07-31 6125c57
ENH: Add noxfiles to the ignore list of autoapi
