### 2021-07-31 #3175 @cheginit
@raoulcollenteur I released new versions for all the packages. Also, the two problematic web services are now back up. So, all the example notebooks now should run without any issues. You can give them a try via [binder](https://mybinder.org/v2/gh/cheginit/hyriver-examples/HEAD?urlpath=lab/tree/notebooks).
