### 2021-08-16 f8b7dad
DOC: Add download stats' badges
