### 2021-08-19 #3175 @kthyng
Hi @cheginit — I have heard by email that @evanleeturner won't have time to do this review after all. I am sorry about the delay in this. I will now recruit a new reviewer.

Hi @arbennett! Would you be able to step in as a reviewer on this submission for JOSS? The review process is already underway, but one reviewer has had to step out, and I am looking for a replacement. I know you've reviewed for JOSS before, but as a reminder you can see more information here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html
