### 2021-08-20 #3175 @arbennett
@kthyng I would be happy to review this! I see the review process has been going for a while so I'll try to complete my review ASAP.  I have one other pending JOSS review at the moment so as soon as I work through that I'll start here.
