### 2021-08-21 74cd893
DOC: Remove the note related to the package name
