+++ docs/source/autoapi/async_retriever/async_retriever/index.rst
@@ -12,6 +12,11 @@
 Module Contents
 ---------------
 
+.. function:: clean_cache(cache_name: Union[Path, str]) -> None
+
+   Remove expired responses from the cache file.
+
+
 .. function:: retrieve(urls: Union[StrOrURL, List[StrOrURL], Tuple[StrOrURL, ...]], read: str, request_kwds: Optional[List[Dict[str, Any]]] = None, request_method: str = 'GET', max_workers: int = 8, cache_name: Optional[Union[Path, str]] = None, family: str = 'both') -> List[Union[str, Dict[str, Any], bytes]]
 
    Send async requests.
