### 2021-08-23 c797061
ENH: Change download stat target link to their respective github repo
