### 2021-08-27 2c01226
DOC: Add API for the new pynhd function
