### 2021-09-04 a88a6bb
ENH: Automatically add notebooks to the gallery
