### 2021-09-04 b7e048b
ENH: Automatically add updated readmes
