+++ Makefile
@@ -46,7 +46,16 @@ lint: ## check style with flake8
 	pre-commit run --all-files
 
 docs: ## generate Sphinx HTML documentation, including API docs
-	for f in $(ls HyRiver-Examples/notebooks/*.ipynb);do echo \{\"path\": \"../../../$f\"\} > docs/source/notebooks/$(basename $f .ipynb).nblink;done
+	for f in $$(ls HyRiver-Examples/notebooks/*.ipynb);do \
+		fname=docs/source/notebooks/$$(basename $$f .ipynb).nblink; \
+		echo \{\"path\": \"../../../$$f\"\} > $$fname; \
+	done
+	for f in pynhd py3dep pygeohydro pydaymet async_retriever pygeoogc pygeoutils;do \
+		end=$$(grep -n Contributing $$f/README.rst | cut -d : -f 1); \
+		end=$$(expr $$end - 2); \
+		fname=docs/source/readme/$$f.rst; \
+		awk 'NR==60, NR=='"$$end"'; NR=='"$$end"' {exit}' $$f/README.rst > $$fname; \
+	done
 	$(MAKE) -C docs clean
 	$(MAKE) -C docs html
 	$(BROWSER) docs/build/html/index.html
