### 2021-09-04 4c6a18e
ENH: Add a new section for each package for decribing their capabilities openjournals/joss-reviews#3175
