+++ docs/source/readme/pygeohydro.rst
@@ -1,0 +1,209 @@
+PyGeoHydro: Retrieve Geospatial Hydrology Data
+----------------------------------------------
+
+.. image:: https://img.shields.io/pypi/v/pygeohydro.svg
+    :target: https://pypi.python.org/pypi/pygeohydro
+    :alt: PyPi
+
+.. image:: https://img.shields.io/conda/vn/conda-forge/pygeohydro.svg
+    :target: https://anaconda.org/conda-forge/pygeohydro
+    :alt: Conda Version
+
+.. image:: https://codecov.io/gh/cheginit/pygeohydro/graph/badge.svg
+    :target: https://codecov.io/gh/cheginit/pygeohydro
+    :alt: CodeCov
+
+.. image:: https://img.shields.io/pypi/pyversions/pygeohydro.svg
+    :target: https://pypi.python.org/pypi/pygeohydro
+    :alt: Python Versions
+
+.. image:: https://pepy.tech/badge/hydrodata
+    :target: https://pepy.tech/project/hydrodata
+    :alt: Downloads
+
+|
+
+.. image:: https://www.codefactor.io/repository/github/cheginit/pygeohydro/badge/main
+    :target: https://www.codefactor.io/repository/github/cheginit/pygeohydro/overview/main
+    :alt: CodeFactor
+
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/psf/black
+    :alt: black
+
+.. image:: https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white
+    :target: https://github.com/pre-commit/pre-commit
+    :alt: pre-commit
+
+.. image:: https://mybinder.org/badge_logo.svg
+    :target: https://mybinder.org/v2/gh/cheginit/HyRiver-examples/main?urlpath=lab/tree/notebooks
+    :alt: Binder
+
+|
+
+Features
+--------
+
+PyGeoHydro (formerly named `hydrodata <https://pypi.org/project/hydrodata>`__) is a part of
+`HyRiver <https://github.com/cheginit/HyRiver>`__ software stack that
+is designed to aid in watershed analysis through web services. This package provides
+access to some of the public web services that offer geospatial hydrology data. It has three
+main modules: ``pygeohydro``, ``plot``, and ``helpers``.
+
+The ``pygeohydro`` module can pull data from the following web services:
+
+* `NWIS <https://nwis.waterdata.usgs.gov/nwis>`__ for daily mean streamflow observations,
+* `NID <https://damsdev.net/>`__ for accessing the National Inventory of Dams in the US,
+* `HCDN 2009 <https://www2.usgs.gov/science/cite-view.php?cite=2932>`__ for identifying sites
+  where human activity affects the natural flow of the watercourse,
+* `NLCD 2019 <https://www.mrlc.gov/>`__ for land cover/land use, imperviousness, imperviousness
+  descriptor, and canopy data,
+* `SSEBop <https://earlywarning.usgs.gov/ssebop/modis/daily>`__ for daily actual
+  evapotranspiration, for both single pixel and gridded data.
+
+Also, it has two other functions:
+
+* ``interactive_map``: Interactive map for exploring NWIS stations within a bounding box.
+* ``cover_statistics``: Categorical statistics of land use/land cover data.
+
+The ``plot`` module includes two main functions:
+
+* ``signatures``: Hydrologic signature graphs.
+* ``cover_legends``: Official NLCD land cover legends for plotting a land cover dataset.
+* ``descriptor_legends``: Color map and legends for plotting a imperviousness descriptor dataset.
+
+The ``helpers`` module includes:
+
+* ``nlcd_helper``: A roughness coefficients lookup table for each land cover and imperviousness
[…]
