+++ docs/source/readme/async_retriever.rst
@@ -1,0 +1,220 @@
+AsyncRetriever: Asynchronous requests with persistent caching
+-------------------------------------------------------------
+
+.. image:: https://img.shields.io/pypi/v/async_retriever.svg
+    :target: https://pypi.python.org/pypi/async_retriever
+    :alt: PyPi
+
+.. image:: https://img.shields.io/conda/vn/conda-forge/async_retriever.svg
+    :target: https://anaconda.org/conda-forge/async_retriever
+    :alt: Conda Version
+
+.. image:: https://codecov.io/gh/cheginit/async_retriever/branch/main/graph/badge.svg
+    :target: https://codecov.io/gh/cheginit/async_retriever
+    :alt: CodeCov
+
+.. image:: https://img.shields.io/pypi/pyversions/async_retriever.svg
+    :target: https://pypi.python.org/pypi/async_retriever
+    :alt: Python Versions
+
+.. image:: https://github.com/cheginit/async_retriever/actions/workflows/test.yml/badge.svg
+    :target: https://github.com/cheginit/async_retriever/actions/workflows/test.yml
+    :alt: Github Actions
+
+|
+
+.. image:: https://img.shields.io/badge/security-bandit-green.svg
+    :target: https://github.com/PyCQA/bandit
+    :alt: Security Status
+
+.. image:: https://www.codefactor.io/repository/github/cheginit/async_retriever/badge
+   :target: https://www.codefactor.io/repository/github/cheginit/async_retriever
+   :alt: CodeFactor
+
+.. image:: https://img.shields.io/badge/code%20style-black-000000.svg
+    :target: https://github.com/psf/black
+    :alt: black
+
+.. image:: https://img.shields.io/badge/pre--commit-enabled-brightgreen?logo=pre-commit&logoColor=white
+    :target: https://github.com/pre-commit/pre-commit
+    :alt: pre-commit
+
+|
+
+Features
+--------
+
+AsyncRetriever is a part of `HyRiver <https://github.com/cheginit/HyRiver>`__ software stack that
+is designed to aid in watershed analysis through web services. This package has only one purpose;
+asynchronously sending requests and retrieving responses as ``text``, ``binary``, or ``json``
+objects. It uses persistent caching to speedup the retrieval even further. Moreover, thanks
+to `nest_asyncio <https://github.com/erdewit/nest_asyncio>`__ you can use this function in
+Jupyter notebooks as well. Although this package is in the HyRiver software stack, it's
+applicable to any HTTP requests.
+
+Please note that since this project is in early development stages, while the provided
+functionalities should be stable, changes in APIs are possible in new releases. But we
+appreciate it if you give this project a try and provide feedback. Contributions are most welcome.
+
+Moreover, requests for additional functionalities can be submitted via
+`issue tracker <https://github.com/cheginit/async_retriever/issues>`__.
+
+Installation
+------------
+
+You can install ``async_retriever`` using ``pip``:
+
+.. code-block:: console
+
+    $ pip install async_retriever
+
+Alternatively, ``async_retriever`` can be installed from the ``conda-forge`` repository
+using `Conda <https://docs.conda.io/en/latest/>`__:
+
+.. code-block:: console
+
+    $ conda install -c conda-forge async_retriever
+
+Quick start
+-----------
+
+AsyncRetriever has one public function: ``retrieve``. By default, this function uses
+``./cache/aiohttp_cache.sqlite`` as the cache file. You can use ``cache_name`` argument
+to customize it. Now, let's see it in action!
+
+As an example for retrieving a ``binary`` response, let's use the DAAC server to get
[…]
