### 2021-09-04 b3c8609
ENH: Add opengraph settings and also disable edit button for docstrings
