### 2021-09-04 740d243
ENH: Install the theme from pip
