### 2021-09-04 171b32f
ENH: Install all sphinx related packages from pip
