### 2021-09-04 82efcae
ENH: Switch to python from conda for readthedocs
