+++ docs/requirements.txt
@@ -1,0 +1,16 @@
+ipykernel
+ipython
+ipywidgets
+jupyter_client
+recommonmark
+pygithub
+nbsphinx
+nbsphinx_link
+sphinx
+sphinx-copybutton
+sphinx-gallery
+sphinx-autoapi
+sphinx-panels
+sphinx-book-theme
+sphinx-book-theme
+sphinxext-opengraph
