+++ .readthedocs.yml
@@ -1,15 +1,9 @@
 version: 2
 
-build:
-    image: latest
-
-conda:
-    environment: ci/requirements/doc.yml
-
 submodules:
   include: all
 
 python:
+  version: "3.8"
   install:
-    - method: pip
-      path: .
+    - requirements: docs/requirements.txt
