+++ ci/requirements/doc.yml
@@ -1,25 +1,0 @@
-name: doc
-channels:
-  - conda-forge
-  - nodefaults
-dependencies:
-  - python=3.8
-  - ipykernel
-  - ipython
-  - ipywidgets
-  - jupyter_client
-  - recommonmark
-  - pygithub
-
-  - pip
-  - pip:
-    - nbsphinx
-    - nbsphinx_link
-    - sphinx
-    - sphinx-copybutton
-    - sphinx-gallery
-    - sphinx-autoapi
-    - sphinx-panels
-    - sphinx-book-theme
-    - sphinx-book-theme
-    - sphinxext-opengraph
