+++ docs/requirements.txt
@@ -2,15 +2,15 @@ ipykernel
 ipython
 ipywidgets
 jupyter_client
-recommonmark
-pygithub
 nbsphinx
 nbsphinx_link
+pygithub
+recommonmark
 sphinx
+sphinx-autoapi
+sphinx-book-theme
+sphinx-book-theme
 sphinx-copybutton
 sphinx-gallery
-sphinx-autoapi
 sphinx-panels
-sphinx-book-theme
-sphinx-book-theme
 sphinxext-opengraph
