### 2021-09-04 8fb58ab
ENH: Use higher quality images for the twitter cards
