### 2021-09-04 5f7e4c7
ENH: Add image alt for visually impaired
