### 2021-09-04 d034b5f
BUG: Fix the link to the og's logo
