### 2021-09-04 3740f1b
DOC: Remove width specification for images
