### 2021-09-13 016b29c
API: Add the new pydaymet functions
