+++ docs/source/autoapi/pydaymet/index.rst
@@ -28,4 +28,3 @@ Package Contents
 
 
 
-
