+++ docs/source/autoapi/pydaymet/pydaymet/index.rst
@@ -12,7 +12,7 @@
 Module Contents
 ---------------
 
-.. py:function:: get_bycoords(coords, dates, crs = DEF_CRS, variables = None, pet = False, region = 'na', time_scale = 'daily')
+.. py:function:: get_bycoords(coords, dates, crs = DEF_CRS, variables = None, region = 'na', time_scale = 'daily', pet = None)
 
    Get point-data from the Daymet database at 1-km resolution.
 
@@ -26,9 +26,6 @@ Module Contents
                 * **variables** (:class:`str` or :class:`list`) -- List of variables to be downloaded. The acceptable variables are:
                   ``tmin``, ``tmax``, ``prcp``, ``srad``, ``vp``, ``swe``, ``dayl``
                   Descriptions can be found `here <https://daymet.ornl.gov/overview>`__.
-                * **pet** (:class:`bool`) -- Whether to compute evapotranspiration based on
-                  `FAO Penman-Monteith equation <http://www.fao.org/3/X0490E/x0490e06.htm>`__.
-                  The default is False
                 * **region** (:class:`str`, *optional*) -- Target region in the US, defaults to ``na``. Acceptable values are:
 
                   * ``na``: Continental North America
@@ -36,6 +33,14 @@ Module Contents
                   * ``pr``: Puerto Rico
                 * **time_scale** (:class:`str`, *optional*) -- Data time scale which can be daily, monthly (monthly summaries),
                   or annual (annual summaries). Defaults to daily.
+                * **pet** (:class:`str`, *optional*) -- Method for computing PET. Supported methods are
+                  ``penman_monteith``, ``priestley_taylor``, ``hargreaves_samani``, and
+                  None (don't compute PET). The ``penman_monteith`` method is based on
+                  :footcite:t:`Allen_1998` assuming that soil heat flux density is zero.
+                  The ``priestley_taylor`` method is based on
+                  :footcite:t:`Priestley_1972` assuming that soil heat flux density is zero.
+                  The ``hargreaves_samani`` method is based on :footcite:t:`Hargreaves_1982`.
+                  Defaults to ``None``.
 
    :returns: :class:`pandas.DataFrame` -- Daily climate data for a location.
 
@@ -44,12 +49,16 @@ Module Contents
    >>> import pydaymet as daymet
    >>> coords = (-1431147.7928, 318483.4618)
    >>> dates = ("2000-01-01", "2000-12-31")
-   >>> clm = daymet.get_bycoords(coords, dates, crs="epsg:3542", pet=True)
+   >>> clm = daymet.get_bycoords(coords, dates, crs="epsg:3542", pet="hargreaves_samani")
    >>> clm["pet (mm/day)"].mean()
-   3.497
+   3.713
+
+   .. rubric:: References
+
+   .. footbibliography::
 
 
-.. py:function:: get_bygeom(geometry, dates, crs = DEF_CRS, variables = None, pet = False, region = 'na', time_scale = 'daily')
+.. py:function:: get_bygeom(geometry, dates, crs = DEF_CRS, variables = None, region = 'na', time_scale = 'daily', pet = None)
 
    Get gridded data from the Daymet database at 1-km resolution.
 
@@ -59,9 +68,6 @@ Module Contents
                 * **variables** (:class:`str` or :class:`list`) -- List of variables to be downloaded. The acceptable variables are:
                   ``tmin``, ``tmax``, ``prcp``, ``srad``, ``vp``, ``swe``, ``dayl``
                   Descriptions can be found `here <https://daymet.ornl.gov/overview>`__.
-                * **pet** (:class:`bool`) -- Whether to compute evapotranspiration based on
-                  `FAO Penman-Monteith equation <http://www.fao.org/3/X0490E/x0490e06.htm>`__.
-                  The default is False
[…]
