### 2021-09-13 bb582f4
ENH: Add a new extention for handling bib
