+++ docs/source/refs.bib
@@ -1,0 +1,34 @@
+@article{Priestley_1972,
+  title={On the assessment of surface heat flux and evaporation using large-scale parameters},
+  author={Priestley, Charles Henry Brian and TAYLOR, Robert Joseph},
+  journal={Monthly weather review},
+  volume={100},
+  number={2},
+  pages={81--92},
+  year={1972}
+}
+
+
+@article{Allen_1998,
+  title={Crop evapotranspiration-Guidelines for computing crop water requirements-FAO Irrigation and drainage paper 56},
+  author={Allen, Richard G and Pereira, Luis S and Raes, Dirk and Smith, Martin and others},
+  journal={Fao, Rome},
+  volume={300},
+  number={9},
+  pages={D05109},
+  year={1998}
+}
+
+@article{Hargreaves_1982,
+	doi = {10.1061/jrcea4.0001390},
+	url = {https://doi.org/10.1061%2Fjrcea4.0001390},
+	year = 1982,
+	month = {sep},
+	publisher = {American Society of Civil Engineers ({ASCE})},
+	volume = {108},
+	number = {3},
+	pages = {225--230},
+	author = {George H. Hargreaves and Zohrab A. Samani},
+	title = {Estimating Potential Evapotranspiration},
+	journal = {Journal of the Irrigation and Drainage Division}
+}
