### 2021-09-14 55c3591
DOC: Add info about pet computation methods
