+++ docs/source/autoapi/py3dep/py3dep/index.rst
@@ -12,21 +12,18 @@
 Module Contents
 ---------------
 
-.. py:function:: deg2mpm(da)
+.. py:function:: elevation_bycoords(coords, crs = DEF_CRS, source = 'airmap')
 
-   Convert slope from degree to meter/meter.
+   Get elevation for a list of coordinates.
 
-
-.. py:function:: elevation_bycoords(coords, crs = DEF_CRS, source = 'tnm')
-
-   Get elevation from Airmap at 1-arc resolution (~30 m) for a list of coordinates.
-
-   :Parameters: * **coords** (:class:`list` of :class:`tuples`) -- Coordinates of target location as list of tuples ``[(x, y), ...]``.
+   :Parameters: * **coords** (:class:`list` of :class:`tuple`) -- Coordinates of target location as list of tuples ``[(x, y), ...]``.
                 * **crs** (:class:`str` or :class:`pyproj.CRS`, *optional*) -- Spatial reference (CRS) of coords, defaults to ``EPSG:4326``.
-                * **source** (:class:`str`, *optional*) -- Data source to be used, default to ``tnm``. Supported sources are
+                * **source** (:class:`str`, *optional*) -- Data source to be used, default to ``airmap``. Supported sources are
                   ``airmap`` (30 m resolution) and ``tnm`` (using The National Map's Bulk Point
-                  Query Service). The ``tnm`` source is more accurate since it uses the highest
-                  available resolution DEM automatically but it is limited to the US.
+                  Query Service with 10 m resolution). The ``tnm`` source is more accurate since it
+                  uses the 1/3 arc-second DEM layer from 3DEP service but it is limited to the US.
+                  It also tends to be slower than the Airmap service and more unstable.
+                  It's recommended to use ``airmap`` unless you need 10-m resolution accuracy.
 
    :returns: :class:`list` of :class:`float` -- Elevation in meter.
 
@@ -37,7 +34,7 @@ Module Contents
 
    This function is intended for getting elevations for a gridded dataset.
 
-   :Parameters: * **xcoords** (:class:`list`) -- List x-coordinates of a a grid.
+   :Parameters: * **xcoords** (:class:`list`) -- List of x-coordinates of a grid.
                 * **ycoords** (:class:`list`) -- List of y-coordinates of a grid.
                 * **crs** (:class:`str`) -- The spatial reference system of the input grid, defaults to ``EPSG:4326``.
                 * **resolution** (:class:`float`) -- The accuracy of the output, defaults to 10 m which is the highest
@@ -46,14 +43,14 @@ Module Contents
                 * **depression_filling** (:class:`bool`, *optional*) -- Fill depressions before sampling using
                   `RichDEM <https://richdem.readthedocs.io/en/latest/>`__ package, defaults to False.
 
-   :returns: :class:`xarray.DataArray` -- An data array with name elevation and the given dim names.
+   :returns: :class:`xarray.DataArray` -- Elevations of the input coordinates as a ``xarray.DataArray``.
 
 
 .. py:function:: get_map(layers, geometry, resolution, geo_crs = DEF_CRS, crs = DEF_CRS)
 
    Access to `3DEP <https://www.usgs.gov/core-science-systems/ngp/3dep>`__ service.
 
-   The 3DEP service has multi-resolution sources so depending on the user
+   The 3DEP service has multi-resolution sources, so depending on the user
    provided resolution the data is resampled on server-side based
    on all the available data sources. The following layers are available:
 
@@ -70,20 +67,16 @@ Module Contents
    - ``Contour 25``
    - ``Contour Smoothed 25``
 
-   :Parameters: * **layers** (:class:`str` or :class:`list`) -- A valid 3DEP layer or a list of them.
[…]
