### 2021-10-12 9c1e430
ENH: Make the order of package listing the same everywhere
