### 2021-10-18 #3175 @arbennett
# Summary

HyRiver is the high level interface to a number of APIs for accessing, analyzing, and visualizing datasets related to hydrologic modeling. It provides a unified interface and is implemented in modern python in a way that is easily scriptable and extensible into larger workflows. HyRiver mainly addresses problems of either setting up large/complex hydrologic models or doing further analysis on primary datasets, and provides ways to access information about a number of important areas from watershed delineations, river networks, and DEMS to metorologic forcing data and land use/cover. I've started this review a few times now and each time get a little distracted while playing with the package itself. I think this is an awesome contribution and can easily see myself using it in my own research.

The paper itself is well written and explains what/where everything is. The only real comment I have is whether it would be worth having a top level installer that installs all of the packages? I see the argument against this too, so it's not really important. I think you've also done a good job explaining the ecosystem of other similar packages, and explained the rationale behind HyRiver. I've also got to commend you on the documentation - it's really nice. I love how many examples you've put together and even made them runnable in binder. I was pretty easily able to go from just `<shift>-<Entering>` my way through the binder to installing locally on my laptop with the same notebooks, and then even to modifying them for a one-off in a new region of interest.

I would say based on all of this, I am happy to recommend accepting this for publication in JOSS! Really nice work here!
